% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-arguments.R
\name{checkDoptions}
\alias{checkDoptions}
\title{Check list D.options.}
\usage{
checkDoptions(n.lag = NULL, n.lead = NULL, kmax = c("k4", "k12"),
  info.crit = c("AIC", "BIC"))
}
\arguments{
\item{n.lag, n.lead}{[\code{NULL} | \code{numeric(1)}]\cr
Have to be "integerish" and > 0.}

\item{kmax}{[\code{NULL} | \code{character(1)}]\cr
One of \code{"k4"} or \code{"k12"}.}

\item{info.crit}{[\code{NULL} | \code{character(1)}]\cr
One of \code{"AIC"} or \code{"BIC"}.}
}
\value{
[\code{list}]. List with the checked and (if necessary)
  converted arguments.

  If one of \code{n.lag} and \code{n.lead} is
  \code{NULL}, only \code{kmax} and \code{info.crit} will be not \code{NULL}.
}
\description{
Checking the list D.options, that is an argument of
\code{\link{cointRegD}}.
}
\examples{
checkDoptions(n.lag = 3, n.lead = 4)
checkDoptions(info.crit = "BIC")
checkDoptions()

# It's not sufficient to include only one of "n.lag" and "n.lead":
checkDoptions(n.lag = 2)

}
\seealso{
Other check: \code{\link{checkObject}},
  \code{\link{checkVars}}
}

